refactor: fixed issues with api for playlists

+++ Mixi.Api/Controllers/PlaylistController.cs
@@ -1,3 +1,5 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Mixi.Api.Modules.Database;
@@ -7,45 +9,69 @@ using Mixi.Shared.Models.Music;
 
 namespace Mixi.Api.Controllers;
 
-
+[Authorize]
 [ApiController]
 [Route("api/playlists")]
 public class PlaylistController : ControllerBase
 {
     
+    private string? GetUserName()
+    {
+        return User.FindFirst(ClaimTypes.Name)?.Value;
+    }
     
     private readonly IPlaylistRepository _playlistRepository;
+    
     
     public PlaylistController(IPlaylistRepository playlistRepository)
     {
         _playlistRepository = playlistRepository;
     }
 
-    [HttpGet("{id}/getPlaylists")]
-    public async Task<ActionResult<IEnumerable<PlaylistDto>>> GetPlaylists(string userId)
+    [HttpGet("getPlaylists")]
+    public async Task<ActionResult<IEnumerable<PlaylistDto>>> GetPlaylists()
     {
         
-        var playlists = _playlistRepository.GetAllPlaylists(userId);
+        var userId = GetUserName();
+
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
+        var playlists = await _playlistRepository.GetAllPlaylists(userId);
         
         return Ok(playlists);
     }
     
-    [HttpPost("{userId}/{id}/getPlaylist")]
-    public async Task<ActionResult<PlaylistDto>> GetPlaylist(int playlistId, string userId)  
+    [HttpGet("{id}/getPlaylist")]
+    public async Task<ActionResult<PlaylistDto>> GetPlaylist(int id)
     {
-      
 
-        var playlist = _playlistRepository.GetPlaylistById(userId, playlistId);
+        var playlistId = id;
+        var userId = GetUserName();
+
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+
+        var playlist = await  _playlistRepository.GetPlaylistById(userId, playlistId);
         
         return Ok(playlist);
     }
 
-    [HttpPost("{id}/createPlaylist")]
-    public async Task<ActionResult<PlaylistDto>> CreatePlaylist([FromBody] CreatePlaylistDto createPlaylistDto, string userId)
+    [HttpPost("createPlaylist")]
+    public async Task<ActionResult<PlaylistDto>> CreatePlaylist([FromBody] CreatePlaylistDto createPlaylistDto)
     {
-      
+      var userId = GetUserName();
+
+      if (userId == null)
+      {
+          return Unauthorized();
+      }
         
-        var newPlaylist = _playlistRepository.CreatePlaylist(createPlaylistDto, userId);;
+        var newPlaylist = await  _playlistRepository.CreatePlaylist(createPlaylistDto, userId);;
         
         return CreatedAtAction(nameof(GetPlaylist), new {id = newPlaylist.Id}, newPlaylist);
     }
@@ -53,27 +79,56 @@ public class PlaylistController : ControllerBase
     [HttpDelete("{id}/deletePlaylist")]
     public async Task<ActionResult> DeletePlaylistbyid(int id)
     {
-        await _playlistRepository.DeletePlaylist(id);
+        
+        var userId = GetUserName();
+
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
+        await _playlistRepository.DeletePlaylist(id, userId);;
         
         return Ok();
     }
 
     [HttpPost("{id}/addTrackPlaylist")]
-    public async Task<ActionResult<PlaylistItemDto>> AddTrackToPlaylist(int id, string userId,
+    public async Task<ActionResult<PlaylistItemDto>> AddTrackToPlaylist(int id, 
         [FromBody] CreatePlaylistItemDto playlistItemDto)
     {
+        var userId = GetUserName();
+
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
         var newPlaylistItemDto = await _playlistRepository.CreatePlaylistItem(playlistItemDto, id, userId);
+
+        if (newPlaylistItemDto == null)
+        {
+            return  NotFound("Playlist not found");
+        }
         
         return Ok(newPlaylistItemDto);
     }
 
-    [HttpDelete("{playlistId}/deleteTrackPlaylist")]
-    public async Task<IActionResult> DeleteTrackFromPlaylist(int playlistId, string userId, int trackId)
+    [HttpDelete("{playlistId}/{trackId}/deleteTrackFromPlaylist")]
+    public async Task<IActionResult> DeleteTrackFromPlaylist(int playlistId, int trackId)
     {
+        
+        var userId = GetUserName();
+        
+        if (userId == null)
+        {
+            return Unauthorized();
+        }
+        
         await _playlistRepository.DeletePlaylistItem(playlistId, userId, trackId);
         
-        return Ok();
+        return NoContent();
     }
+
